+++ example/lib/main_text_one_line.dart
@@ -14,38 +14,81 @@ class Demo extends StatelessWidget {
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: const [
-            IntrinsicHeight(child: TextOneLine("ELLIPSIS", style: TextStyle(fontWeight: FontWeight.bold))),
+            //
+            // -------------------
+            //
             SizedBox(height: 10),
-            IntrinsicWidth(child: TextOneLine("short: Lorem ipsum")),
-            SizedBox(height: 10),
-            IntrinsicWidth(child: IntrinsicHeight(child: TextOneLine("medium: Lorem ipsum dolor sit amet, consectetur adipiscing elit"))),
-            SizedBox(height: 10),
+            Text("Text with ELLIPSIS, maxLines: 1, softWrap: false",
+                maxLines: 1,
+                softWrap: false,
+                overflow: TextOverflow.ellipsis,
+                style: TextStyle(fontWeight: FontWeight.bold)),
+            SizedBox(height: 5),
+            Text("short: Lorem ipsum",
+                maxLines: 1, softWrap: false, overflow: TextOverflow.ellipsis),
+            SizedBox(height: 5),
+            Text(
+                "long: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+                maxLines: 1,
+                softWrap: false,
+                overflow: TextOverflow.ellipsis),
+            SizedBox(height: 5),
+            Text(
+                "long without spaces: This is a verylongtextwithoutanyspacesthatdemonstratestheproblemwegetwhentryingtousethenativetextwidget.",
+                maxLines: 1,
+                softWrap: false,
+                overflow: TextOverflow.ellipsis),
+            SizedBox(height: 40),
+            //
+            // -------------------
+            //
+            TextOneLine("TextOneLine with ELLIPSIS", style: TextStyle(fontWeight: FontWeight.bold)),
+            SizedBox(height: 5),
+            TextOneLine("short: Lorem ipsum"),
+            SizedBox(height: 5),
             TextOneLine(
                 "long: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."),
-            SizedBox(height: 20),
+            SizedBox(height: 5),
+            TextOneLine(
+                "long without spaces: This is a verylongtextwithoutanyspacesthatdemonstratestheproblemwegetwhentryingtousethenativetextwidget.",
+                overflow: TextOverflow.ellipsis),
+            SizedBox(height: 40),
             //
-            TextOneLine("FADE", overflow: TextOverflow.fade, style: TextStyle(fontWeight: FontWeight.bold)),
-            SizedBox(height: 10),
+            // -------------------
+            //
+            TextOneLine("TextOneLine with FADE",
+                overflow: TextOverflow.fade, style: TextStyle(fontWeight: FontWeight.bold)),
+            SizedBox(height: 5),
             TextOneLine("short: Lorem ipsum", overflow: TextOverflow.fade),
-            SizedBox(height: 10),
-            TextOneLine("medium: Lorem ipsum dolor sit amet, consectetur adipiscing elit", overflow: TextOverflow.fade),
-            SizedBox(height: 10),
+            SizedBox(height: 5),
             TextOneLine(
               "long: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
               overflow: TextOverflow.fade,
             ),
-            SizedBox(height: 20),
+            SizedBox(height: 5),
+            TextOneLine(
+                "long without spaces: This is a verylongtextwithoutanyspacesthatdemonstratestheproblemwegetwhentryingtousethenativetextwidget.",
+                overflow: TextOverflow.fade),
+            SizedBox(height: 40),
             //
-            TextOneLine("CLIP", overflow: TextOverflow.clip, style: TextStyle(fontWeight: FontWeight.bold)),
-            SizedBox(height: 10),
+            // -------------------
+            //
+            TextOneLine("TextOneLine with CLIP",
+                overflow: TextOverflow.clip, style: TextStyle(fontWeight: FontWeight.bold)),
+            SizedBox(height: 5),
             TextOneLine("short: Lorem ipsum", overflow: TextOverflow.clip),
-            SizedBox(height: 10),
-            TextOneLine("medium: Lorem ipsum dolor sit amet, consectetur adipiscing elit", overflow: TextOverflow.clip),
-            SizedBox(height: 10),
+            SizedBox(height: 5),
             TextOneLine(
               "long: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
               overflow: TextOverflow.clip,
             ),
+            SizedBox(height: 5),
+            TextOneLine(
+                "long without spaces: This is a verylongtextwithoutanyspacesthatdemonstratestheproblemwegetwhentryingtousethenativetextwidget.",
+                overflow: TextOverflow.clip),
+            //
+            // -------------------
+            //
           ],
         ),
       ),